chore: organize TS config and package.json scripts better
With future CI/CD in mind. But it also needed some refinements, so that both dev and prod build work correctly.

Also, I did my best in removing as much from `tsconfig.server.json` as possible.

Now the stuff should work, of course only when there's the node version expressed in .nvmrc available.

The nodemon was replaced by tsx, which seems to be more recent and recommended.